Make new message to update the curve rather than completely overwriting it.
The old overwriting methods have been renamed to indicate that they are creating a new curve. create is the term to indicate a new plot will be created.

New methods have been added to update curves. update is the term used to indicate a curve is being updated.

Original Commit Rev: 1168 | 2013-09-17 23:10:19 -0500 (Tue, 17 Sep 2013)